add usePapi hook
